Выносит в отдельные модули другие части разметки
