+++ src/main.js
@@ -1,11 +1,8 @@
 import {createFilterTemplate} from './components/filter.js';
-import {createFormEditTemplate} from './components/form-edit.js';
 import {createFormAddTemplate} from './components/form-add.js';
 import {createSiteMenuTemplate} from './components/site-menu.js';
-import {createCardTemplate} from './components/card.js';
 import {createInfoTripTemplate} from './components/info-trip.js';
-
-const CARD_COUNT = 3;
+import {createTripDaysTemplate} from './components/trip-days.js';
 
 const render = (container, template, place) => {
   container.insertAdjacentHTML(place, template);
@@ -24,8 +21,5 @@ const sitePageMainElement = document.querySelector(`.page-main`);
 const siteTripEventsElement = sitePageMainElement.querySelector(`.trip-events`);
 
 render(siteTripEventsElement, createFormAddTemplate(), `beforeend`);
-render(siteTripEventsElement, createFormEditTemplate(), `beforeend`);
 
-const getCardsTemplate = () => new Array(CARD_COUNT).fill(``).map(createCardTemplate).join(``);
-
-render(siteTripEventsElement, getCardsTemplate(), `beforeend`);
+render(siteTripEventsElement, createTripDaysTemplate(), `beforeend`);
